% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_coordinates.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calcular a Distancia entre Dois Pontos}
\usage{
calculate_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Coordenada x do primeiro ponto.}

\item{y1}{Coordenada y do primeiro ponto.}

\item{x2}{Coordenada x do segundo ponto.}

\item{y2}{Coordenada y do segundo ponto.}
}
\value{
Distancia Euclidiana entre os pontos.
}
\description{
Calcular a Distancia entre Dois Pontos
}
