% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gain.R
\name{calculate_gain}
\alias{calculate_gain}
\title{Analisar a Eficiencia do Empacotamento Hexagonal de Formas Geometricas}
\usage{
calculate_gain(
  D,
  S,
  shape = "circle",
  rect_height = NULL,
  xlim = 20,
  exibir_resultados = FALSE,
  exibir_plot = TRUE
)
}
\arguments{
\item{D}{Diametro das formas geometricas (em unidades).}

\item{S}{Tamanho da area disponivel para empacotamento (em unidades).}

\item{shape}{Forma geometrica a ser empacotada. Pode ser "circle" para circulos ou "semi_rect" para
semicerculos unidos por retangulo. Default e "circle".}

\item{rect_height}{Altura do retangulo que une os semicerculos. Necessario apenas quando shape e "semi_rect".}

\item{xlim}{Valor maximo do plot do grafico}

\item{exibir_resultados}{Booleano que determina se os resultados detalhados devem ser exibidos.
Default e FALSE.}

\item{exibir_plot}{Plot o grafico de ganho em relação a altura quando \code{semi_rect}.
Default e TRUE.}
}
\value{
Retorna uma lista com os seguintes componentes:
\describe{
\item{total_shapes}{Numero total de formas que cabem na area utilizando empacotamento hexagonal.}
\item{odd_rows}{Numero de fileiras impares no empacotamento.}
\item{even_rows}{Numero de fileiras pares no empacotamento.}
\item{shapes_per_row_odd}{Numero de formas por fileira impar.}
\item{shapes_per_row_even}{Numero de formas por fileira par.}
\item{num_rows}{Numero total de fileiras no empacotamento.}
\item{rect_packing_shapes}{Numero de formas que caberiam na area utilizando empacotamento retangular.}
\item{increase_percentage}{Porcentagem de aumento na capacidade com empacotamento hexagonal.}
\item{altura_total_empacotamento}{Altura total do empacotamento hexagonal.}
\item{altura_superior_empacotamento}{Altura superior do empacotamento hexagonal.}
}
}
\description{
Esta funcao calcula o ganho percentual na capacidade de empacotamento de formas geometricas
(circulos ou obrounds) utilizando o metodo hexagonal em comparacao com o empacotamento retangular.
}
\examples{

\dontrun{
# Exemplo 1: Analisar empacotamento hexagonal de circulos com diametro 1 em uma area de 1000000 unidades
resultado_circulos <- calculate_gain(D = 1, S = 10000, shape = "circle", exibir_resultados = TRUE)

# Exemplo 2: Analisar empacotamento hexagonal de obrounds com diametro 1, altura do retangulo 2 em uma area de 1000000 unidades
resultado_obrounds <- calculate_gain(D = 1, S = 10000, shape = "semi_rect", rect_height = 2, exibir_resultados = TRUE)
}

}
