% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_aux_plot}
\alias{circ_aux_plot}
\title{Plotar Arvores e Circulo}
\usage{
circ_aux_plot(trees, trees_circle, central_point, radius, title)
}
\arguments{
\item{trees}{Dataframe contendo as coordenadas x e y das arvores.}

\item{trees_circle}{Dataframe contendo as coordenadas x e y das arvores dentro do circulo.}

\item{central_point}{Dataframe com as coordenadas x e y do ponto central.}

\item{radius}{Raio do circulo (em metros).}

\item{title}{Titulo do grafico.}
}
\value{
Nenhum. A funcao gera um grafico.
}
\description{
Esta funcao plota as arvores e um circulo com base nos parametros fornecidos.
}
