% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_central_point}
\alias{circ_central_point}
\title{Obter Ponto Central de Acordo com a Metodologia}
\usage{
circ_central_point(methodology, spacing_x, spacing_y)

circ_central_point(methodology, spacing_x, spacing_y)
}
\arguments{
\item{methodology}{Metodologia a ser usada. Opcoes: "central", "vertical", "horizontal", "quadrante".}

\item{spacing_x}{Espacamento entre as arvores na direcao X (em metros).}

\item{spacing_y}{Espacamento entre as arvores na direcao Y (em metros).}
}
\value{
Retorna um dataframe com as coordenadas x e y do ponto central.

Retorna um dataframe com as coordenadas x e y do ponto central.
}
\description{
Esta funcao retorna o ponto central baseado na metodologia selecionada.

Esta funcao retorna o ponto central baseado na metodologia selecionada.
}
