% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_count}
\alias{circ_count}
\title{Contar Numero de Arvores por Linha a Partir do Ponto Central}
\usage{
circ_count(trees_circle, central_point, spacing_y)
}
\arguments{
\item{trees_circle}{Dataframe contendo as coordenadas x e y das arvores dentro do circulo.}

\item{central_point}{Dataframe com as coordenadas x e y do ponto central.}

\item{spacing_y}{Espacamento entre as arvores na direcao Y (em metros).}
}
\value{
Retorna um dataframe com o numero de arvores por linha e a distancia vertical da linha ao ponto central.
}
\description{
Esta funcao conta o numero de arvores em cada linha horizontal dentro de um circulo a partir do ponto central.
}
