% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_density.R
\name{circ_density}
\alias{circ_density}
\title{Analise de Densidade de arvores em Funcao do Raio e Metodologia}
\usage{
circ_density(
  radius_values,
  spacing_x,
  spacing_y,
  methodologies = c("central", "vertical", "horizontal", "quadrante"),
  observed_tph = NULL
)
}
\arguments{
\item{radius_values}{Vetor de raios (em metros) para os quais calcular o numero de arvores.}

\item{spacing_x}{Espacamento entre as arvores na direcao X (em metros).}

\item{spacing_y}{Espacamento entre as arvores na direcao Y (em metros).}

\item{methodologies}{Vetor com as metodologias a serem utilizadas. Opcoes: "central", "vertical", "horizontal", "quadrante".}

\item{observed_tph}{Numero observado de arvores por hectare (calculado com base no espacamento real).}
}
\value{
Retorna um data frame com o raio, metodologia, numero de arvores no circulo,
area do circulo, densidade estimada (arvores/ha) e erro percentual.
}
\description{
Esta funcao calcula o numero de arvores que cabem em circulos de diferentes raios,
utilizando quatro metodologias diferentes. Em seguida, extrapola o numero de arvores
para 1 hectare e calcula o erro percentual em relacao ao valor observado.
}
\examples{

# Exemplo

radius_values <- seq(1, 30, by = 0.1)
results <- circ_density(
  radius_values = radius_values,
  spacing_x = 3,
  spacing_y = 3,
  methodologies = c("central", "vertical", "horizontal", "quadrante")
)

# Visualizar os primeiros resultados

head(results)

}
