% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_filter}
\alias{circ_filter}
\title{Filtrar Arvores Dentro de um Circulo}
\usage{
circ_filter(trees, central_point, radius)

circ_filter(trees, central_point, radius)
}
\arguments{
\item{trees}{Dataframe contendo as coordenadas x e y das arvores.}

\item{central_point}{Dataframe com as coordenadas x e y do ponto central.}

\item{radius}{Raio do circulo (em metros).}
}
\value{
Retorna um dataframe com as arvores que estao dentro do circulo.

Retorna um dataframe com as arvores que estao dentro do circulo.
}
\description{
Esta funcao filtra as arvores que estao dentro de um circulo definido por um ponto central e um raio.

Esta funcao filtra as arvores que estao dentro de um circulo definido por um ponto central e um raio.
}
