% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_min_radius_for_error.R
\name{circ_min_radius_for_error}
\alias{circ_min_radius_for_error}
\title{Encontrar o Menor Raio onde o Erro Permanece dentro do Intervalo Especificado}
\usage{
circ_min_radius_for_error(results, error_bands)
}
\arguments{
\item{results}{Data frame contendo os resultados da função \code{circ_density}.}

\item{error_bands}{Vetor numérico com os valores de erro para os quais deseja encontrar os raios mínimos.}
}
\value{
Data frame com as colunas: \code{Methodology}, \code{Error_Band}, \code{Min_Radius}.
}
\description{
Esta função recebe os resultados da função \code{circ_density} e um vetor de valores de erro (error bands).
Para cada metodologia e cada valor de erro especificado, a função identifica o menor raio
após o último raio onde o erro percentual ultrapassa o valor do erro especificado (positivo ou negativo).
}
\examples{

# Define parameters
radius_values <- seq(5, 30, by = 0.1)
spacing_x <- 3  # Spacing between trees in the X direction (in meters)
spacing_y <- 2  # Spacing between trees in the Y direction (in meters)

# Run analysis
results <- circ_density(
  radius_values = radius_values,
  spacing_x = spacing_x,
  spacing_y = spacing_y,
  methodologies = c("central", "vertical", "horizontal", "quadrant")
)
error_bands <- c(2.5, 5, 10)
min_radius_df <- circ_min_radius_for_error(results, error_bands)
print(min_radius_df)

}
