% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_plot_density_error.R
\name{circ_plot_density_error}
\alias{circ_plot_density_error}
\title{Plotar Erro Percentual vs. Raio para Diferentes Metodologias}
\usage{
circ_plot_density_error(results, methodology = NULL, colors = NULL)
}
\arguments{
\item{results}{Data frame contendo os resultados da funcao \code{circ_density_analysis}.}

\item{methodology}{Metodologia a ser plotada. Pode ser "central", "vertical",
"horizontal", "quadrante" ou \code{NULL}. Se for \code{NULL}, todas as metodologias
serao plotadas em um unico grafico.}

\item{colors}{Vetor nomeado de cores personalizadas para cada metodologia (opcional).
Exemplo: \code{colors = c("central" = "blue", "vertical" = "green")}.}
}
\value{
Gera graficos do Erro Percentual vs. Raio.
}
\description{
Esta funcao gera graficos do Erro Percentual em funcao do Raio,
com base nos resultados da funcao \code{circ_density_analysis}.
}
\examples{

# Essa funcao necessita executar anteriormente  `circ_density_analysis`

# Preparando os dados

radius_values <- seq(1, 30, by = 0.1)

results <- circ_density(
  radius_values = radius_values,
  spacing_x = 3,
  spacing_y = 3,
  methodologies = c("central", "vertical", "horizontal", "quadrante")
)

# Visualizar os primeiros resultados

head(results)

# Para plotar o Erro Percentual vs. Raio para todas as metodologias:

circ_plot_density_error(results)

# Para plotar todas as metodologias com cores personalizadas:

circ_plot_density_error(
  results,
  colors = c("central" = "#F17CB0", "vertical" = "#60BD68",
             "horizontal" = "#FAA43A", "quadrante" = "#5DA5DA")
)

# Para plotar apenas a metodologia "central":

circ_plot_density_error(results, methodology = "central")

# Outros exemplos

circ_plot_density_error(results, methodology = "central", colors = c(central = '#F17CB0'))
circ_plot_density_error(results, methodology = "vertical", colors = c(vertical = '#60BD68'))
circ_plot_density_error(results, methodology = "horizontal")
circ_plot_density_error(results, methodology = "quadrante")

}
