% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_plot_erros_radius.R
\name{circ_plot_erros_radius}
\alias{circ_plot_erros_radius}
\title{Plotar Erro Amostral e Raio em Função do Tamanho da Amostra}
\usage{
circ_plot_erros_radius(results)
}
\arguments{
\item{results}{Data.frame resultante da função \code{circ_erros_radius}.}
}
\value{
Gera dois gráficos: Erro vs. Tamanho da Amostra e Raio vs. Tamanho da Amostra.

\dontrun{
}
\description{
Esta função gera gráficos do erro amostral e do raio da parcela circular em função do tamanho da amostra.
}
\examples{
# Supondo que você já tenha executado a função `circ_erros_radius` e obtido `results`

\dontrun{

circ_plot_erros_radius(results)

}

}
