% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_radius.R
\name{circ_radius}
\alias{circ_radius}
\title{Calcular Raio Necessario com os Parametros Fornecidos}
\usage{
circ_radius(
  spacing_x = NULL,
  spacing_y = NULL,
  n = NULL,
  methodology = "central",
  area = NULL,
  side_x = NULL,
  side_y = NULL,
  units = "ha",
  option = "inferior"
)
}
\arguments{
\item{spacing_x}{Espacamento entre as arvores na direcao X (em metros). Opcional se \code{n} for fornecido.}

\item{spacing_y}{Espacamento entre as arvores na direcao Y (em metros). Opcional se \code{n} for fornecido.}

\item{n}{Numero de arvores desejado. Opcional se \code{area}, \code{side_x} e \code{side_y} forem fornecidos.}

\item{methodology}{Metodologia a ser usada. Opcoes: "central", "vertical", "horizontal", "quadrante". Padrao e "central".}

\item{area}{Area da parcela. Pode ser fornecida em hectares ("ha") ou metros quadrados ("m2"). Opcional se \code{side_x} e \code{side_y} forem fornecidos.}

\item{side_x}{Comprimento do lado X da parcela (em metros). Opcional se \code{area} for fornecida.}

\item{side_y}{Comprimento do lado Y da parcela (em metros). Opcional se \code{area} for fornecida.}

\item{units}{Unidade de medida da area. Pode ser "ha" para hectares ou "m2" para metros quadrados. Padrao e "ha".}

\item{option}{Opcao de arredondamento para o numero de arvores. Pode ser "inferior" para arredondar para baixo, "superior" para arredondar para cima ou "exato" para manter o valor exato. Padrao e "inferior".}
}
\value{
Retorna o raio (em metros) necessario para incluir o numero desejado de arvores.
}
\description{
Esta funcao calcula o raio necessario para incluir um numero especifico de arvores com base nos parametros fornecidos.
}
\examples{
# Exemplo 1: Calcular o raio para 1666 arvores
circ_radius(spacing_x = 2, spacing_y = 2, n = 1666, methodology = "central")

# Exemplo 2: Calcular o raio para 2 hectares com metodologia quadrante
circ_radius(spacing_x = 3, area = 2, methodology = "quadrante", units = "ha", option = "superior")

}
