% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_radius_for_trees}
\alias{circ_radius_for_trees}
\title{Calcular Raio Necessario para Incluir Numero Desejado de Arvores}
\usage{
circ_radius_for_trees(trees, central_point, desired_tree_count)
}
\arguments{
\item{trees}{Dataframe contendo as coordenadas x e y das arvores.}

\item{central_point}{Dataframe com as coordenadas x e y do ponto central.}

\item{desired_tree_count}{Numero desejado de arvores a serem incluidas no circulo.}
}
\value{
Retorna o raio (em metros) necessario para incluir o numero desejado de arvores.
}
\description{
Esta funcao calcula o raio necessario para incluir um numero especifico de arvores a partir de um ponto central.
}
