% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_aux.R
\name{circ_walk_scheme}
\alias{circ_walk_scheme}
\title{Gerar Esquema de Caminhamento como Tibble com Line_number}
\usage{
circ_walk_scheme(line_counter, central_point, spacing_y)
}
\arguments{
\item{line_counter}{Dataframe contendo as informacoes de cada linha, incluindo \code{Line}, \code{Vertical_distance} e \code{Number_of_trees}.}

\item{central_point}{Dataframe com as coordenadas x e y do ponto central.}

\item{spacing_y}{Espacamento entre as arvores na direcao Y (em metros).}
}
\value{
Retorna um tibble com as colunas \code{Line_number}, \code{Direction}, \code{Vertical_distance}, \code{Number_of_trees} e \code{Message}, representando o esquema de caminhamento.
}
\description{
Esta funcao gera um esquema de caminhamento baseado no contador de linhas, ponto central e espaçamento vertical. O esquema inclui a direcao (Norte, Sul ou Central), o numero da linha, a distancia vertical e uma mensagem descritiva.
}
