% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestdeck-package.R
\docType{package}
\name{forestdeck-package}
\alias{forestdeck}
\alias{forestdeck-package}
\title{forestdeck: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Gustavo Mourao \email{gustavo-mourao@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
