% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_aux_coordinates.R
\name{hex_circle_coordinates}
\alias{hex_circle_coordinates}
\title{Criar Coordenadas de um Circulo}
\usage{
hex_circle_coordinates(x0, y0, r, npoints = 100)
}
\arguments{
\item{x0}{Coordenada x do centro.}

\item{y0}{Coordenada y do centro.}

\item{r}{Raio do circulo.}

\item{npoints}{Numero de pontos para desenhar o circulo.}
}
\value{
Data frame com as coordenadas x e y do circulo.
}
\description{
Criar Coordenadas de um Circulo
}
