% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_height.R
\name{hex_height}
\alias{hex_height}
\title{Calcular a Altura entre Linhas no Empacotamento Hexagonal}
\usage{
hex_height(diameter, shape = "circle", rect_height = NULL)
}
\arguments{
\item{diameter}{Diametro das figuras geometricas. Este parametro e essencial para
determinar o tamanho das formas a serem empacotadas.}

\item{shape}{Forma das figuras a serem empacotadas. Pode ser "circle"
para empacotamento de circulos ou "semi_rect" para empacotamento de
semicirculos unidos por um retangulo.}

\item{rect_height}{Altura do retangulo que une os semicirculos. Este parametro
e necessario apenas quando shape e "semi_rect".}
}
\value{
Retorna a distancia vertical calculada necessaria para o empacotamento das figuras.
}
\description{
Esta funcao calcula a distancia vertical necessaria para organizar figuras geometricas
em um padrao de empacotamento hexagonal, que e o metodo mais eficiente para
empacotar circulos sem sobreposicao. O empacotamento hexagonal maximiza o uso do espaco,
reduzindo o desperdicio e aumentando a densidade de empacotamento em comparacao
com outros metodos, como o empacotamento quadrado.
}
\details{
\subsection{Empacotamento Hexagonal para Circulos}{

No caso de empacotamento de circulos, o empacotamento hexagonal posiciona cada circulo
de forma que cada um seja cercado por seis outros circulos, formando uma estrutura em
formato de colmeia.
}

\subsection{Uso com Semicirculos e Retangulos}{

Quando a forma especificada e \code{semi_rect}, a funcao considera uma geometria composta por
dois semicirculos unidos por um retangulo. Neste caso, os semicirculos utilizam o raio
definido e o retangulo e acrescentado entre eles com uma altura especificada. Por exemplo,
um semicirculo com raio 3 e um retangulo com altura 2 resultaria em uma forma geometrica
conhecida como \strong{obround} (tambem chamada de \strong{estadio} ou \strong{discóide} em português).
Esta forma teria uma largura de 6 unidades (diametro dos semicirculos) e uma altura de 5
unidades (raio + altura do retangulo).
}
}
\examples{

\dontrun{
# Exemplo 1: Empacotamento hexagonal de circulos com raio 5
altura_circulo <- hex_height(diameter = 3, shape = "circle")
print(altura_circulo)

# Exemplo 2: Empacotamento de semicirculos unidos por um retangulo com raio 3 e altura do retangulo 2
altura_semi_rect <- hex_height(diameter = 2, shape = "semi_rect", rect_height = 1)
print(altura_semi_rect)

}

}
