% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_height_circle.R
\name{hex_height_circle}
\alias{hex_height_circle}
\title{Calcular a Altura no Empacotamento Hexagonal de Circulos}
\usage{
hex_height_circle(diameter)
}
\arguments{
\item{diameter}{Diametro dos circulos geometricos. Este parametro e essencial para
determinar o tamanho das formas a serem empacotadas.}
}
\value{
Retorna a distancia vertical calculada necessaria para o empacotamento hexagonal de circulos.
}
\description{
Esta funcao calcula a distancia vertical necessaria para organizar circulos em um padrao
de empacotamento hexagonal, que e o metodo mais eficiente para empacotar circulos sem sobreposicao.
O empacotamento hexagonal maximiza o uso do espaco, reduzindo o desperdicio e aumentando a
densidade de empacotamento em comparacao com outros metodos, como o empacotamento quadrado.
}
\details{
\subsection{Empacotamento Hexagonal para Circulos}{

No empacotamento hexagonal de circulos, cada circulo e posicionado de forma que seja cercado
por seis outros circulos, formando uma estrutura similar a uma colmeia. Este arranjo resulta
em uma maxima densidade de empacotamento de aproximadamente 90,7\%, o que significa que 90,7\%
da area esta ocupada pelos circulos, minimizando o espaco vazio.
}
}
\examples{

\dontrun{
# Exemplo: Calcular a altura para circulos com diametro 10
altura <- hex_height_circle(diameter = 10)
print(altura)
}

}
