% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot.R
\name{hex_plot}
\alias{hex_plot}
\title{Plotar o Arranjo Hexagonal das Formas}
\usage{
hex_plot(area_size, diameter, shape = "circle", rect_height = NULL)
}
\arguments{
\item{area_size}{Tamanho da area disponivel para empacotamento (por exemplo, em unidades).}

\item{diameter}{Diametro das figuras geometricas (circulos ou semicirculos). Este parametro e essencial para
determinar o tamanho das formas a serem empacotadas.}

\item{shape}{Forma das figuras a serem empacotadas. Pode ser "circle"
para empacotamento de circulos ou "semi_rect" para empacotamento de semicirculos unidos por retangulo.}

\item{rect_height}{Altura do retangulo que une os semicirculos. Este parametro e necessario apenas quando shape e "semi_rect".}
}
\value{
Retorna um objeto ggplot representando o arranjo hexagonal das formas dentro da area definida.
}
\description{
Esta funcao plota o arranjo hexagonal das formas geometricas (circulos ou obrounds) dentro de uma area
definida. Utiliza o metodo de empacotamento hexagonal para organizar as formas de maneira eficiente,
maximizando a densidade de empacotamento e minimizando o desperdicio de espaco.
}
\details{
\subsection{Empacotamento Hexagonal para Circulos}{

No empacotamento hexagonal de circulos, cada circulo e cercado por seis outros circulos, formando
uma estrutura semelhante a uma colmeia. Este arranjo resulta em uma densidade de empacotamento
maxima de aproximadamente 90,7\%, otimizando o uso do espaco disponivel.
}

\subsection{Empacotamento Hexagonal Adaptado para Obrounds}{

Quando a forma especificada e \code{"semi_rect"}, a funcao considera uma geometria composta por
dois semicirculos unidos por um retangulo, formando uma \strong{obround} (tambem conhecida como
\strong{estadio} ou \strong{discoide} em portugues). Este arranjo combina a eficiencia do empacotamento
de semicirculos com a versatilidade do retangulo, resultando em uma forma que pode ser utilizada
em diversas aplicacões geometricas.
}
}
\examples{

\dontrun{
# Exemplo 1: Plotar o arranjo hexagonal de circulos com diametro 10 em uma area de 100 unidades
plot_circles <- hex_plot(area_size = 100, diameter = 10, shape = "circle")
print(plot_circles)

# Exemplo 2: Plotar o arranjo hexagonal adaptado para obrounds com diametro 6 e altura do retangulo 2 em uma area de 100000 unidades
plot_obrounds <- hex_plot(area_size = 100000, diameter = 6, shape = "semi_rect", rect_height = 2)
print(plot_obrounds)

}

}
