% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_aux_coordinates.R
\name{hex_semi_rect_coordinates}
\alias{hex_semi_rect_coordinates}
\title{Criar Coordenadas de um Semicirculo com Retangulo}
\usage{
hex_semi_rect_coordinates(x0, y0, r, rect_height, npoints = 100)
}
\arguments{
\item{x0}{Coordenada x do centro.}

\item{y0}{Coordenada y do centro.}

\item{r}{Raio do semicirculo.}

\item{rect_height}{Altura do retangulo.}

\item{npoints}{Numero de pontos para desenhar o semicirculo.}
}
\value{
Data frame com as coordenadas x e y do semicirculo com retangulo.
}
\description{
Criar Coordenadas de um Semicirculo com Retangulo
}
